nomad_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.13
requests>=2.25,<3.0
django-cors-headers